+++ Common/src/main/java/net/merchantpug/apugli/power/factory/ModifyEnchantmentLevelPowerFactory.java
@@ -2,7 +2,6 @@ package net.merchantpug.apugli.power.factory;
 
 import io.github.apace100.calio.data.SerializableData;
 import io.github.apace100.calio.data.SerializableDataTypes;
-import net.merchantpug.apugli.access.ItemStackAccess;
 import net.merchantpug.apugli.platform.Services;
 import net.merchantpug.apugli.util.ComparableItemStack;
 import net.minecraft.core.Registry;
@@ -17,7 +16,8 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.Level;
 
-import java.util.*;
+import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
 public interface ModifyEnchantmentLevelPowerFactory<P> extends ValueModifyingPowerFactory<P> {
@@ -62,7 +62,7 @@ public interface ModifyEnchantmentLevelPowerFactory<P> extends ValueModifyingPow
     }
 
     default ListTag generateEnchantments(ListTag enchants, ItemStack self) {
-        Entity entity = ((ItemStackAccess) (Object) self).getEntity();
+        Entity entity = Services.PLATFORM.getEntityFromItemStack(self);
 
         if(!(entity instanceof LivingEntity living)) return enchants;
 
@@ -93,7 +93,7 @@ public interface ModifyEnchantmentLevelPowerFactory<P> extends ValueModifyingPow
     }
 
     default ListTag getEnchantments(ItemStack self, ListTag originalTag) {
-        Entity entity = ((ItemStackAccess) (Object) self).getEntity();
+        Entity entity = Services.PLATFORM.getEntityFromItemStack(self);
         if (entity instanceof LivingEntity living && getEntityItemEnchants().containsKey(entity.getStringUUID())) {
             ConcurrentHashMap<ComparableItemStack, ListTag> itemEnchants = getEntityItemEnchants().get(entity.getStringUUID());
             ComparableItemStack comparableStack = new ComparableItemStack(self.copy());
@@ -121,7 +121,7 @@ public interface ModifyEnchantmentLevelPowerFactory<P> extends ValueModifyingPow
     }
 
     default int getItemEnchantmentLevel(Enchantment enchantment, ItemStack self) {
-        Entity entity = ((ItemStackAccess) (Object) self).getEntity();
+        Entity entity = Services.PLATFORM.getEntityFromItemStack(self);
         if (entity instanceof LivingEntity living && getEntityItemEnchants().containsKey(living.getStringUUID())) {
             ConcurrentHashMap<ComparableItemStack, ListTag> itemEnchants = getEntityItemEnchants().get(entity.getStringUUID());
             ComparableItemStack comparableStack = new ComparableItemStack(self);