+++ Forge/src/main/java/net/merchantpug/apugli/capability/entity/HitsOnTargetCapability.java
@@ -1,7 +1,6 @@
-package net.merchantpug.apugli.capability;
+package net.merchantpug.apugli.capability.entity;
 
 import net.merchantpug.apugli.network.ApugliPacketHandler;
-import net.merchantpug.apugli.network.c2s.UpdateKeysPressedPacket;
 import net.merchantpug.apugli.network.s2c.SyncHitsOnTargetCapabilityPacket;
 import net.merchantpug.apugli.util.ApugliConfigs;
 import net.minecraft.core.Direction;